+++ packages/connect/e2e/test-yarn-install.sh
@@ -16,10 +16,10 @@ npm init -y
 touch yarn.lock
 
 # install connect package
-yarn add @trezor/connect
+yarn add @trezor/connect@"$1"
 # prepare minimal typescript implementation
-echo import TrezorConnect from \"@trezor/connect\" > index.ts
+echo import TrezorConnect from \"@trezor/connect\" >index.ts
 
 # compile with typescript
-yarn add typescript@5.3.2
-yarn tsc ./index.ts --types node --esModuleInterop
+yarn add typescript@5.5.4
+yarn tsc ./index.ts --types node --types w3c-web-usb --esModuleInterop