+++ packages/product-components/src/components/CoinLogo/coins.ts
@@ -1,0 +1,28 @@
+import { NetworkSymbol } from '@suite-common/wallet-config';
+
+export const COINS: Record<NetworkSymbol, string> = {
+    ada: require('../../images/coins/ada.svg'),
+    bch: require('../../images/coins/bch.svg'),
+    bnb: require('../../images/coins/bnb.svg'),
+    btc: require('../../images/coins/btc.svg'),
+    btg: require('../../images/coins/btg.svg'),
+    dash: require('../../images/coins/dash.svg'),
+    dgb: require('../../images/coins/dgb.svg'),
+    doge: require('../../images/coins/doge.svg'),
+    dsol: require('../../images/coins/dsol.svg'),
+    etc: require('../../images/coins/etc.svg'),
+    eth: require('../../images/coins/eth.svg'),
+    ltc: require('../../images/coins/ltc.svg'),
+    matic: require('../../images/coins/matic.svg'),
+    nmc: require('../../images/coins/nmc.svg'),
+    regtest: require('../../images/coins/btc_test.svg'),
+    sol: require('../../images/coins/sol.svg'),
+    tada: require('../../images/coins/tada.svg'),
+    test: require('../../images/coins/btc_test.svg'),
+    thol: require('../../images/coins/thol.svg'),
+    tsep: require('../../images/coins/tsep.svg'),
+    txrp: require('../../images/coins/txrp.svg'),
+    vtc: require('../../images/coins/vtc.svg'),
+    xrp: require('../../images/coins/xrp.svg'),
+    zec: require('../../images/coins/zec.svg'),
+};