+++ suite-native/intl/src/en.ts
@@ -261,6 +261,23 @@ export const en = {
             button: 'Enter PIN again',
         },
     },
+    moduleConnectPopup: {
+        title: 'Trezor Connect Mobile',
+        callback: 'Callback',
+        confirm: 'Confirm',
+        areYouSureMessage: 'Are you sure you want to continue?\nMake sure you trust the source.',
+        connectionStatus: {
+            loading: 'Loading...',
+            discoveryRunning: 'Discovery running, please wait...',
+        },
+        errors: {
+            deviceNotConnected: 'Device not connected.',
+            invalidCallback: 'Invalid callback URL',
+        },
+        bottomSheets: {
+            confirmOnDeviceMessage: 'Go to your device and verify the details of the operation.',
+        },
+    },
     moduleDevice: {
         incompatibleFirmwareModalAppendix: {
             title: 'Follow these steps',
@@ -787,6 +804,9 @@ export const en = {
             button: 'Enter passphrase on Trezor',
             title: 'Continue on Trezor',
             subtitle: 'Enter your passphrase on your Trezor',
+        },
+        noPassphrase: {
+            button: 'No passphrase',
         },
         loading: {
             title: 'Checking passphrase wallet for balances & transactions',