+++ suite-native/module-connect-popup/src/index.ts
@@ -1,0 +1,3 @@
+export * from './screens/ConnectPopupScreen';
+export * from './hooks/useConnectPopupNavigation';
+export * from './hooks/useIsConnectPopupOpened';